Update README.md
Added logs for today, starting a new section called "LOG_2024", after an 8-year hiatus after writing only a partial log for a single day in May, 2014. I will try to keep this log as comprehensive as possible and up to date. It is my new location for the logging of my "Daily Acts".